refactor(ssg): split pager tests to stay under the file line limit
Move the pagination unit tests into a sibling module so the new source file stays within 350 lines.
